Drive the streaming decision through real inputs in five session-lifecycle tests
Rewrites five tests in test_response_runner_session_lifecycle.py to stop
patching mindroom.response_runner.should_use_streaming: the off case is now
driven by the enable_streaming runtime input and the on case by requester
presence in the fake client's synced room cache, so the real decision logic
and its wiring stay covered.

Also extracts the _bot/_config/_envelope/_target/_plain_request/_noop_typing
scaffolding from test_response_runner_focused.py into
tests/response_runner_helpers.py so test_locked_turn_delivery.py no longer
imports from a test module.
