+++ tests/test_response_runner_focused.py
@@ -9,7 +9,6 @@ orchestrator/bot boot, so shrinking ``response_runner.py`` has a safety net.
 from __future__ import annotations
 
 import asyncio
-from contextlib import asynccontextmanager
 from dataclasses import replace
 from typing import TYPE_CHECKING
 from unittest.mock import AsyncMock, MagicMock, patch
@@ -19,21 +18,13 @@ import pytest
 from mindroom import background_tasks as background_tasks_module
 from mindroom import response_runner
 from mindroom.background_tasks import wait_for_background_tasks
-from mindroom.bot import AgentBot
-from mindroom.config.agent import AgentConfig
-from mindroom.config.auth import AuthorizationConfig
-from mindroom.config.main import Config
-from mindroom.config.models import ModelConfig
 from mindroom.constants import STREAM_STATUS_KEY, STREAM_STATUS_PENDING
 from mindroom.conversation_resolver import ConversationResolver, MessageContext
 from mindroom.delivery_gateway import DeliveryGateway
 from mindroom.final_delivery import FinalDeliveryOutcome, StreamTransportOutcome
 from mindroom.history.turn_recorder import TurnRecorder
-from mindroom.hooks import MessageEnvelope
 from mindroom.logging_config import get_logger
 from mindroom.matrix.cache import ThreadHistoryResult
-from mindroom.matrix.users import AgentMatrixUser
-from mindroom.message_target import MessageTarget
 from mindroom.post_response_effects import PostResponseEffectsDeps, ResponseOutcome, apply_post_response_effects
 from mindroom.response_attempt import ResponseAttemptDeps, ResponseAttemptRequest, ResponseAttemptRunner
 from mindroom.response_lifecycle import ResponseLifecycleCoordinator
@@ -43,18 +34,19 @@ from mindroom.stop import StopManager
 from mindroom.streaming import StreamingDeliveryError
 from mindroom.turn_policy import PreparedDispatch
 from tests.conftest import (
-    TEST_PASSWORD,
-    bind_runtime_paths,
-    install_runtime_cache_support,
     make_matrix_client_mock,
-    message_origin,
     patch_response_runner_module,
     replace_response_runner_deps,
     request_envelope,
-    runtime_paths_for,
-    test_runtime_paths,
     unwrap_extracted_collaborator,
-    wrap_extracted_collaborators,
+)
+from tests.response_runner_helpers import (
+    _bot,
+    _config,
+    _envelope,
+    _noop_typing,
+    _plain_request,
+    _target,
 )
 
 if TYPE_CHECKING:
@@ -63,57 +55,8 @@ if TYPE_CHECKING:
 
     from nio import AsyncClient
 
-
-def _config(tmp_path: Path) -> Config:
-    return bind_runtime_paths(
-        Config(
-            agents={"general": AgentConfig(display_name="General", rooms=["!room:localhost"])},
-            teams={},
-            models={"default": ModelConfig(provider="openai", id="test-model")},
-            authorization=AuthorizationConfig(default_room_access=True),
-        ),
-        test_runtime_paths(tmp_path),
-    )
-
-
-def _bot(tmp_path: Path) -> AgentBot:
-    config = _config(tmp_path)
-    agent_user = AgentMatrixUser(
-        agent_name="general",
-        password=TEST_PASSWORD,
-        display_name="General",
-        user_id="@mindroom_general:localhost",
-    )
-    bot = AgentBot(agent_user, tmp_path, config, runtime_paths_for(config), rooms=["!room:localhost"])
-    bot.client = make_matrix_client_mock(user_id="@mindroom_general:localhost")
-    install_runtime_cache_support(bot)
-    wrap_extracted_collaborators(bot)
-    return bot
-
-
-def _target(*, thread_id: str | None = None, reply_to_event_id: str = "$event") -> MessageTarget:
-    return MessageTarget.resolve(
-        room_id="!room:localhost",
-        thread_id=thread_id,
-        reply_to_event_id=reply_to_event_id,
-        room_mode=thread_id is None,
-    )
-
-
-def _envelope(target: MessageTarget, *, source_event_id: str = "$event") -> MessageEnvelope:
-    return MessageEnvelope(
-        source_event_id=source_event_id,
-        room_id="!room:localhost",
-        target=target,
-        requester_id="@user:localhost",
-        sender_id="@user:localhost",
-        body="hello",
-        attachment_ids=(),
-        mentioned_agents=(),
-        agent_name="general",
-        source_kind="message",
-        origin=message_origin(sender_id="@user:localhost", requester_id="@user:localhost", source_kind="message"),
-    )
+    from mindroom.hooks import MessageEnvelope
+    from mindroom.message_target import MessageTarget
 
 
 def _preparation(target: MessageTarget, envelope: MessageEnvelope) -> ResponsePayloadPreparation:
@@ -145,15 +88,6 @@ def _preparation(target: MessageTarget, envelope: MessageEnvelope) -> ResponsePa
     )
 
 
-def _plain_request(target: MessageTarget, *, source_event_id: str = "$event") -> ResponseRequest:
-    return ResponseRequest(
-        thread_history=[],
-        prompt="hello",
-        user_id="@user:localhost",
-        response_envelope=_envelope(target, source_event_id=source_event_id),
-    )
-
-
 def _completed_outcome(event_id: str = "$response", body: str = "ok") -> FinalDeliveryOutcome:
     return FinalDeliveryOutcome(
         terminal_status="completed",
@@ -162,11 +96,6 @@ def _completed_outcome(event_id: str = "$response", body: str = "ok") -> FinalDe
         final_visible_body=body,
         delivery_kind="sent",
     )
-
-
-@asynccontextmanager
-async def _noop_typing(*_args: object, **_kwargs: object) -> AsyncIterator[None]:
-    yield
 
 
 class RecordingStopManager(StopManager):
